Moved @tryghost/update-check-service to services folder
fix https://linear.app/ghost/issue/ENG-2039/move-tryghostupdate-check-service-to-services-folder

- this service is already self-contained and can be moved into the
  services folder, which means we can also remove the package that was supporting it
- minor other changes just to fix paths
