+++ ghost/core/package.json
@@ -158,7 +158,6 @@
     "@tryghost/string": "0.2.12",
     "@tryghost/tiers": "0.0.0",
     "@tryghost/tpl": "0.1.32",
-    "@tryghost/update-check-service": "0.0.0",
     "@tryghost/url-utils": "4.4.8",
     "@tryghost/validator": "0.2.14",
     "@tryghost/verification-trigger": "0.0.0",
