+++ ghost/core/test/unit/server/services/update-check.test.js
@@ -5,7 +5,7 @@ const crypto = require('crypto');
 const assert = require('assert/strict');
 
 const logging = require('@tryghost/logging');
-const UpdateCheckService = require('../lib/UpdateCheckService');
+const UpdateCheckService = require('../../../../core/server/services/update-check/UpdateCheckService');
 
 describe('Update Check', function () {
     const internal = {context: {internal: true}};
