--- ghost/update-check-service/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@tryghost/update-check-service",
-  "private": true,
-  "version": "0.0.0",
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "./lib/UpdateCheckService.js",
-  "exports": "./lib/UpdateCheckService.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --check-coverage --reporter text --reporter cobertura -- mocha --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "lib"
-  ],
-  "devDependencies": {
-    "c8": "8.0.1",
-    "mocha": "10.2.0",
-    "should": "13.2.3",
-    "sinon": "15.2.0",
-    "uuid": "9.0.1"
-  },
-  "dependencies": {
-    "@tryghost/debug": "0.1.32",
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/tpl": "0.1.32",
-    "lodash": "4.17.21",
-    "moment": "2.24.0"
-  }
-}
